+++ games/naruto-roguelite/src/data/catalog/index.js
@@ -16,5 +16,6 @@ export * from './characters.js';
 export * from './enemies.js';
 export * from './bosses.js';
 export * from './missions.js';
+export * from './factions.js';
 export * from './regions.js';
 export * from './items.js';